Rename components.py to component_interactions.py
