+++ hikari/interactions/__init__.py
@@ -25,4 +25,4 @@ from __future__ import annotations
 
 from hikari.interactions.base_interactions import *
 from hikari.interactions.command_interactions import *
-from hikari.interactions.components import *
+from hikari.interactions.component_interactions import *
