+++ hikari/interactions/__init__.pyi
@@ -3,4 +3,4 @@
 
 from hikari.interactions.base_interactions import *
 from hikari.interactions.command_interactions import *
-from hikari.interactions.components import *
+from hikari.interactions.component_interactions import *
